refactor dashboard widgets
